+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_reserved_argument.py
@@ -1,10 +1,14 @@
+from contextlib import suppress
+
 import pytest
 
 from wemake_python_styleguide.constants import SPECIAL_ARGUMENT_NAMES_WHITELIST
 from wemake_python_styleguide.violations.naming import (
     ReservedArgumentNameViolation,
 )
-from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+from wemake_python_styleguide.visitors.ast.naming.validation import (
+    WrongNameVisitor,
+)
 
 # Correct:
 
@@ -23,47 +27,28 @@ class Test(object):
 
 correct_function_template = 'def function({0}, different): ...'
 
-# Wrong:
-
-wrong_method_template = """
-class Test(object):
-    def method(different, {0}):
-        ...
-"""
-
-wrong_classmethod_template = """
-class Test(object):
-    @classmethod
-    def method(different, {0}):
-        ...
-"""
-
-wrong_function_template = 'def function(different, {0}): ...'
-
 
 @pytest.mark.parametrize('argument', SPECIAL_ARGUMENT_NAMES_WHITELIST)
-@pytest.mark.parametrize('code', [
-    wrong_method_template,
-    wrong_classmethod_template,
-    wrong_function_template,
-])
 def test_reserved_argument_name(
     assert_errors,
     assert_error_text,
     parse_ast_tree,
     default_options,
-    code,
+    naming_template,
     mode,
     argument,
 ):
     """Ensures that special names for arguments are restricted."""
-    tree = parse_ast_tree(mode(code.format(argument)))
+    with suppress(SyntaxError):
+        # We use `suppress` here because some fixutres
+        # have duplicated `self` or `cls` arguments.
+        tree = parse_ast_tree(mode(naming_template.format(argument)))
 
-    visitor = WrongNameVisitor(default_options, tree=tree)
-    visitor.run()
+        visitor = WrongNameVisitor(default_options, tree=tree)
+        visitor.run()
 
-    assert_errors(visitor, [ReservedArgumentNameViolation])
-    assert_error_text(visitor, argument)
+        assert_errors(visitor, [ReservedArgumentNameViolation])
+        assert_error_text(visitor, argument)
 
 
 @pytest.mark.parametrize('argument', SPECIAL_ARGUMENT_NAMES_WHITELIST)