+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_trailing_underscore.py
@@ -3,26 +3,49 @@ import pytest
 from wemake_python_styleguide.violations.naming import (
     TrailingUnderscoreViolation,
 )
-from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+from wemake_python_styleguide.visitors.ast.naming.validation import (
+    WrongNameVisitor,
+)
+
+wrong_name = 'my_variable_'
 
 
 def test_wrong_trailing_underscore(
     assert_errors,
     assert_error_text,
     parse_ast_tree,
-    naming_template,
+    own_naming_template,
     default_options,
     mode,
 ):
     """Test names with trailing underscores."""
-    wrong_name = 'my_variable_'
-    tree = parse_ast_tree(mode(naming_template.format(wrong_name)))
+    tree = parse_ast_tree(
+        mode(own_naming_template.format(wrong_name)),
+    )
 
     visitor = WrongNameVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [TrailingUnderscoreViolation])
     assert_error_text(visitor, wrong_name)
+
+
+def test_foreign_trailing_underscore(
+    assert_errors,
+    parse_ast_tree,
+    foreign_naming_template,
+    default_options,
+    mode,
+):
+    """Test names with trailing underscores."""
+    tree = parse_ast_tree(
+        mode(foreign_naming_template.format(wrong_name)),
+    )
+
+    visitor = WrongNameVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
 
 
 @pytest.mark.parametrize('correct_name', [