+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_wrong_unused_name.py
@@ -3,7 +3,9 @@ import pytest
 from wemake_python_styleguide.violations.naming import (
     WrongUnusedVariableNameViolation,
 )
-from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+from wemake_python_styleguide.visitors.ast.naming.validation import (
+    WrongNameVisitor,
+)
 
 
 @pytest.mark.parametrize('wrong_name', [