+++ wemake_python_styleguide/logic/naming/builtins.py
@@ -4,17 +4,13 @@ import keyword
 
 from typing_extensions import Final
 
-from wemake_python_styleguide.constants import UNUSED_PLACEHOLDER
+from wemake_python_styleguide.constants import BUILTINS_WHITELIST
 from wemake_python_styleguide.logic.naming.access import is_magic, is_unused
-
-_BUILTINS_WHITELIST: Final = frozenset((
-    UNUSED_PLACEHOLDER,
-))
 
 _BUILTINS: Final = frozenset((
     builtin[0]
     for builtin in inspect.getmembers(builtins)
-    if builtin[0] not in _BUILTINS_WHITELIST
+    if builtin[0] not in BUILTINS_WHITELIST
 ))
 
 _ALL_BUILTINS: Final = frozenset((