+++ wemake_python_styleguide/logic/tree/variables.py
@@ -1,6 +1,8 @@
 import ast
 from typing import Union
 
+from wemake_python_styleguide.logic import nodes
+
 _VarDefinition = Union[ast.AST, ast.expr]
 _LocalVariable = Union[ast.Name, ast.ExceptHandler]
 
@@ -10,6 +12,21 @@ def get_variable_name(node: _LocalVariable) -> str:
     if isinstance(node, ast.Name):
         return node.id
     return getattr(node, 'name', '')
+
+
+def does_shadow_builtin(node: ast.AST) -> bool:
+    """
+    We allow attributes and class-level builtin overrides.
+
+    Like: ``self.list = []`` or ``def map(self, function):``
+
+    Why?
+    Because they cannot harm you since they do not shadow the real builtin.
+    """
+    return (
+        not isinstance(node, ast.Attribute) and
+        not isinstance(nodes.get_context(node), ast.ClassDef)
+    )
 
 
 def is_valid_block_variable_definition(node: _VarDefinition) -> bool: